fix git over ssh when the default http/https ports are not used.
Internally the SSH and the http/https servers always bind to ports 22 and
80/443 respectively. If the GITLAB_PORT variable is set to any other
port, then cloning over SSH would not work.

This is because internally gitlab-shell tries to communicate with the
gitlab server over http interface. In gitlab-shell configured with
the GITLAB_PORT value the conmmunication will effective fail and thereby
the cloning operation would return an error.

To fix this, we make gitlab-shell communicate with the http server
directly with the default http/https ports since this is all internal.

Prior to this update, cloning repositories over ssh protocol would not
work when the container is started in the following way:

```bash
docker run -i -t --rm \
  -p 8888:80 -p 2222:22 \
  -e "GITLAB_PORT=8888" -e "GITLAB_SHELL_SSH_PORT=2222" \
  sameersbn/gitlab
```