run gitlab init script as git user Closes #103 #74
When the /etc/init.d/gitlab script tries to switch the user to the `git` user, selinux steps in and blocks it. To avoid this we simple run the script as the `git` user using the `sudo -u git -H` command prefix.

In short the image should work with selinux enabled.