init: uid/gid adaption
To this day, Docker does not support mapping user/group ids from the
host to the container. Hence, all files created on the mounted volumes
are owned by git:git (1000:1000). On the host that uid/gid possibly
belongs to a different user/group.

With this change the container will adapt the uid/gid of git in
/etc/passwd and /etc/group to the environment variables USERMAP_UID and
USERMAP_GID if specified.

In addition all files in /home/git and /var/log/gitlab have to be
adapted to reflect those changes. This can be a time consuming process
which can be improved by committing the container after the first run.

Idea from: http://chapeau.freevariable.com/2014/08/docker-uid.html